chore(assets): rename assets
naming rule: prefix_name_suffix
