+++ src/pages/Welcome/Welcome.tsx
@@ -1,12 +1,11 @@
+import { useEffect } from 'react';
+import { useNavigate } from 'react-router-dom';
 import styles from './Welcome.module.scss';
-
-import WelcomeBanner from '../../assets/welcome-banner.webp';
-import GreenTransparentLogo from '../../assets/mapu-logo/green-transparent.svg';
 import useRegisterStore from '../../stores/registerStore';
 import { RegisterStatus } from '../../types/RegisterStatus';
-import { useNavigate } from 'react-router-dom';
-import { useEffect } from 'react';
 import AuthContainer from '../../components/login/AuthContainer';
+import WelcomeBanner from '../../assets/welcome/img_welcome_banner.webp';
+import GreenTransparentLogo from '../../assets/logo/ico_logo_transparent_green.svg';
 
 const Welcome = () => {
   const navigate = useNavigate();
